+++ src/template.rs
@@ -159,7 +159,6 @@ pub fn default_solution() -> R<String> {
 		r#"// 💖 Hi, thanks for using ICIE! 💖
 // 🔧 To use a custom code template, press Ctrl+Shift+P and select "ICIE Template configure" from the list 🔧
 // 📝 If you spot any bugs or miss any features, create an issue at https://github.com/pustaczek/icie/issues 📝
-// 🎉 ... or just ask me about anything via Discord at https://discord.gg/ffSEhtb 🎉
 "#,
 		false,
 		"",